refactor (ontology): create separate file for class def.
